refactor: reorganize utility functions
